step-4-dynamic-ui/ai-backend/requirements.txt
# AI Backend Dependencies - Step 2 - Python 3.13 Compatible
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.25.2

# LLM and AI
langchain>=0.0.350
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10

# For Ollama
ollama>=0.1.7

# For OpenRouter
openai>=1.6.1

# Include RAG service dependencies
qdrant-client==1.14.3
sentence-transformers>=2.2.2
numpy>=1.24.3
pandas>=2.0.3
loguru>=0.7.2
aiofiles>=23.2.0

# Observability
langfuse>=2.0.0